+++ src/attitude-utils.hpp
@@ -5,6 +5,7 @@
 #include <vector>
 
 #include "serialize-helpers.hpp"
+#include "decimal.hpp"
 
 namespace lost {
 
@@ -12,58 +13,58 @@ namespace lost {
 // to Quaterinon, and another storing as Quaternion and converting to Euler. But abstract classes
 // make everything more annoying, because you need vectors of pointers...ugh!
 
-/// A two dimensional vector with floating point components
+/// A two dimensional vector with decimaling point components
 struct Vec2 {
-    float x;
-    float y;
+    decimal x;
+    decimal y;
 
-    float Magnitude() const;
-    float MagnitudeSq() const;
+    decimal Magnitude() const;
+    decimal MagnitudeSq() const;
 
     Vec2 Normalize() const;
 
-    float operator*(const Vec2 &) const;
-    Vec2 operator*(const float &) const;
+    decimal operator*(const Vec2 &) const;
+    Vec2 operator*(const decimal &) const;
     Vec2 operator-(const Vec2 &) const;
     Vec2 operator+(const Vec2 &) const;
 };
 
 class Mat3; // define above so we can use in Vec3 class
 
-/// Three dimensional vector with floating point components
+/// Three dimensional vector with decimaling point components
 class Vec3 {
 public:
-    float x;
-    float y;
-    float z;
+    decimal x;
+    decimal y;
+    decimal z;
 
-    float Magnitude() const;
-    float MagnitudeSq() const;
+    decimal Magnitude() const;
+    decimal MagnitudeSq() const;
     Vec3 Normalize() const;
 
-    float operator*(const Vec3 &) const;
-    Vec3 operator*(const float &) const;
+    decimal operator*(const Vec3 &) const;
+    Vec3 operator*(const decimal &) const;
     Vec3 operator*(const Mat3 &) const;
     Vec3 operator-(const Vec3 &) const;
     Vec3 CrossProduct(const Vec3 &) const;
     Mat3 OuterProduct(const Vec3 &) const;
 };
 
-/// 3x3 vector with floating point components
+/// 3x3 vector with decimaling point components
 class Mat3 {
 public:
-    float x[9];
+    decimal x[9];
 
-    float At(int i, int j) const;
+    decimal At(int i, int j) const;
     Mat3 operator+(const Mat3 &) const;
     Mat3 operator*(const Mat3 &) const;
     Vec3 operator*(const Vec3 &) const;
-    Mat3 operator*(const float &) const;
+    Mat3 operator*(const decimal &) const;
     Mat3 Transpose() const;
     Vec3 Column(int) const;
     Vec3 Row(int) const;
-    float Trace() const;
-    float Det() const;
+    decimal Trace() const;
+    decimal Det() const;
     Mat3 Inverse() const;
 };
 
@@ -72,8 +73,8 @@ extern const Mat3 kIdentityMat3;
 void SerializeVec3(SerializeContext *, const Vec3 &);
 Vec3 DeserializeVec3(DeserializeContext *des);
 
-float Distance(const Vec2 &, const Vec2 &);
-float Distance(const Vec3 &, const Vec3 &);
+decimal Distance(const Vec2 &, const Vec2 &);
+decimal Distance(const Vec3 &, const Vec3 &);
 
 /**
  * A "human-readable" way to represent a 3d rotation or orientation.
@@ -82,15 +83,15 @@ float Distance(const Vec3 &, const Vec3 &);
  */
 class EulerAngles {
 public:
-    EulerAngles(float ra, float de, float roll)
+    EulerAngles(decimal ra, decimal de, decimal roll)
         : ra(ra), de(de), roll(roll) { };
 
     /// Right ascension. How far we yaw left. Yaw is performed first.
-    float ra;
+    decimal ra;
     /// Declination. How far we pitch up (or down if negative). Pitch is performed second, after yaw.
-    float de;
+    decimal de;
     /// How far we roll counterclockwise. Roll is performed last (after yaw and pitch).
-    float roll;
+    decimal roll;
 };
 
 /// A quaternion is a common way to represent a 3d rotation.
@@ -98,9 +99,9 @@ class Quaternion {
 public:
     Quaternion() = default;
     explicit Quaternion(const Vec3 &);
-    Quaternion(const Vec3 &, float);
+    Quaternion(const Vec3 &, decimal);
 
-    Quaternion(float real, float i, float j, float k)
+    Quaternion(decimal real, decimal i, decimal j, decimal k)
         : real(real), i(i), j(j), k(k) { };
 
     Quaternion operator*(const Quaternion &other) const;
@@ -108,19 +109,19 @@ public:
     Vec3 Vector() const;
     void SetVector(const Vec3 &);
     Vec3 Rotate(const Vec3 &) const;
-    float Angle() const;
+    decimal Angle() const;
     /// Returns the smallest angle that can be used to represent the rotation represented by the
     /// quaternion. I.e, min(Angle, 2pi-Angle).
-    float SmallestAngle() const;
-    void SetAngle(float);
+    decimal SmallestAngle() const;
+    void SetAngle(decimal);
     EulerAngles ToSpherical() const;
-    bool IsUnit(float tolerance) const;
+    bool IsUnit(decimal tolerance) const;
     Quaternion Canonicalize() const;
 
-    float real;
-    float i;
-    float j;
-    float k;
+    decimal real;
+    decimal i;
+    decimal j;
+    decimal k;
 };
 
 //
@@ -165,23 +166,23 @@ Quaternion DCMToQuaternion(const Mat3 &);
 /// Return a quaternion that will reorient the coordinate axes so that the x-axis points at the given
 /// right ascension and declination, then roll the coordinate axes counterclockwise (i.e., the stars
 /// will appear to rotate clockwise). This is an "improper" z-y'-x' Euler rotation.
-Quaternion SphericalToQuaternion(float ra, float dec, float roll);
+Quaternion SphericalToQuaternion(decimal ra, decimal dec, decimal roll);
 
 /// returns unit vector
-Vec3 SphericalToSpatial(float ra, float de);
-void SpatialToSpherical(const Vec3 &, float *ra, float *de);
+Vec3 SphericalToSpatial(decimal ra, decimal de);
+void SpatialToSpherical(const Vec3 &, decimal *ra, decimal *de);
 /// angle between two vectors, using dot product and magnitude division
-float Angle(const Vec3 &, const Vec3 &);
+decimal Angle(const Vec3 &, const Vec3 &);
 /// angle between two vectors, /assuming/ that they are already unit length
-float AngleUnit(const Vec3 &, const Vec3 &);
+decimal AngleUnit(const Vec3 &, const Vec3 &);
 
-float RadToDeg(float);
-float DegToRad(float);
-float RadToArcSec(float);
-float ArcSecToRad(float);
-/// Given a float, find it "modulo" another float, in the true mathematical sense (not remainder).
+decimal RadToDeg(decimal);
+decimal DegToRad(decimal);
+decimal RadToArcSec(decimal);
+decimal ArcSecToRad(decimal);
+/// Given a decimal, find it "modulo" another decimal, in the true mathematical sense (not remainder).
 /// Always returns something in [0,mod) Eg -0.8 mod 0.6 = 0.4
-float FloatModulo(float x, float mod);
+decimal FloatModulo(decimal x, decimal mod);
 
 // TODO: quaternion and euler angle conversion, conversion between ascension/declination to rec9tu
