+++ src/camera.hpp
@@ -13,16 +13,16 @@ public:
     /**
      * @param xCenter,yCenter The "principal point" of the camera. In an ideal camera, just half the resolution, but physical cameras often have a bit of offset.
      */
-    Camera(float focalLength,
-           float xCenter, float yCenter,
+    Camera(decimal focalLength,
+           decimal xCenter, decimal yCenter,
            int xResolution, int yResolution)
         : focalLength(focalLength),
           xCenter(xCenter), yCenter(yCenter),
           xResolution(xResolution), yResolution(yResolution) {};
 
-    Camera(float focalLength, int xResolution, int yResolution)
+    Camera(decimal focalLength, int xResolution, int yResolution)
         : Camera(focalLength,
-                 xResolution / (float) 2.0, yResolution / (float) 2.0,
+                 xResolution / (decimal) 2.0, yResolution / (decimal) 2.0,
                  xResolution, yResolution) {};
 
     Vec2 SpatialToCamera(const Vec3 &) const;
@@ -30,7 +30,7 @@ public:
 
     // converts from a 2d point in the camera sensor to right ascension and declination relative to
     // the center of the camera.
-    // void CoordinateAngles(const Vec2 &vector, float *ra, float *de) const;
+    // void CoordinateAngles(const Vec2 &vector, decimal *ra, decimal *de) const;
 
     bool InSensor(const Vec2 &vector) const;
 
@@ -39,20 +39,20 @@ public:
     /// Height of the sensor in pixels
     int YResolution() const { return yResolution; };
     /// Focal length in pixels
-    float FocalLength() const { return focalLength; };
+    decimal FocalLength() const { return focalLength; };
     /// Horizontal field of view in radians
-    float Fov() const;
+    decimal Fov() const;
 
-    void SetFocalLength(float focalLength) { this->focalLength = focalLength; }
+    void SetFocalLength(decimal focalLength) { this->focalLength = focalLength; }
 
 private:
     // TODO: distortion
-    float focalLength;
-    float xCenter; float yCenter;
+    decimal focalLength;
+    decimal xCenter; decimal yCenter;
     int xResolution; int yResolution;
 };
 
-float FovToFocalLength(float xFov, float xResolution);
+decimal FovToFocalLength(decimal xFov, decimal xResolution);
 
 }
